requirements.txt
absl-py==0.9.0
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cropyble==1.2.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1
django-crispy-forms==1.9.2
django-filter==2.3.0
djangorestframework==3.11.1
entrypoints==0.3
gast==0.3.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
imageai==2.1.5
imageio==2.8.0
imutils==0.5.3
incf.countryutils==1.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
isort==5.5.2
jedi==0.17.2
Jinja2==2.11.2
joblib==0.15.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
jupyterlab==2.2.5
jupyterlab-server==1.2.0
jupyterthemes==0.20.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lesscpy==0.14.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mysqlclient==2.0.1
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.1.3
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.3.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plotly==4.9.0
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytesseract==0.3.5
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.3.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.0.0
tifffile==2020.5.25
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
